requirements.txt
boto3
langchain
numpy
pandas
python-dotenv
streamlit
streamlit-local-storage